Introduce `GraknClient` interface
