+++ grakn-client/src/main/java/ai/grakn/client/GraknClient.java
@@ -14,101 +14,33 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
+ *
  */
+
 package ai.grakn.client;
 
-import ai.grakn.GraknTxType;
 import ai.grakn.Keyspace;
 import ai.grakn.graql.Query;
-import ai.grakn.util.REST;
 import ai.grakn.util.SimpleURI;
-import com.sun.jersey.api.client.Client;
-import com.sun.jersey.api.client.ClientResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-import javax.ws.rs.core.Response.Status.Family;
-import javax.ws.rs.core.UriBuilder;
-import java.net.URI;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
-
-import static ai.grakn.util.REST.Request.Graql.ALLOW_MULTIPLE_QUERIES;
-import static ai.grakn.util.REST.Request.Graql.EXECUTE_WITH_INFERENCE;
-import static ai.grakn.util.REST.Request.Graql.LOADING_DATA;
-import static ai.grakn.util.REST.Request.Graql.TX_TYPE;
-import static ai.grakn.util.REST.Response.ContentType.APPLICATION_JSON;
 
 /**
  * Grakn http client. Extend this for more http endpoint.
  *
  * @author Domenico Corapi
  */
-public class GraknClient {
+public interface GraknClient {
+    int CONNECT_TIMEOUT_MS = 30 * 1000;
+    int DEFAULT_MAX_RETRY = 3;
 
-    public static final int CONNECT_TIMEOUT_MS = 30 * 1000;
-    public static final int DEFAULT_MAX_RETRY = 3;
-    private final Logger LOG = LoggerFactory.getLogger(GraknClient.class);
-
-    private final Client client;
-    private final SimpleURI uri;
-
-    public GraknClient(SimpleURI url)  {
-        this.client = Client.create();
-        this.client.setConnectTimeout(CONNECT_TIMEOUT_MS);
-        this.client.setReadTimeout(CONNECT_TIMEOUT_MS * 2);
-        this.uri = url;
+    static GraknClient of(SimpleURI url) {
+        return new GraknClientImpl(url);
     }
 
-    public List<QueryResponse> graqlExecute(List<Query<?>> queryList, Keyspace keyspace)
-            throws GraknClientException {
-        LOG.debug("Sending query list size {} to keyspace {}", queryList.size(), keyspace);
+    List<QueryResponse> graqlExecute(List<Query<?>> queryList, Keyspace keyspace)
+            throws GraknClientException;
 
-        String body = queryList.stream().map(Object::toString).reduce("; ", String::concat).substring(2);
-        URI fullURI = UriBuilder.fromUri(uri.toURI())
-                .path(REST.resolveTemplate(REST.WebPath.KEYSPACE_GRAQL, keyspace.getValue()))
-                .queryParam(ALLOW_MULTIPLE_QUERIES, true)
-                .queryParam(EXECUTE_WITH_INFERENCE, false) //Making inference true could lead to non-deterministic loading of data
-                .queryParam(LOADING_DATA, true) //Skip serialising responses for the sake of efficiency
-                .queryParam(TX_TYPE, GraknTxType.BATCH)
-                .build();
-        ClientResponse response = client.resource(fullURI)
-                .accept(APPLICATION_JSON)
-                .post(ClientResponse.class, body);
-        try {
-            Response.StatusType status = response.getStatusInfo();
-            String entity = response.getEntity(String.class);
-            if (!status.getFamily().equals(Family.SUCCESSFUL)) {
-                String queries = queryList.stream().map(Object::toString).collect(Collectors.joining("\n"));
-                throw new GraknClientException("Failed graqlExecute. Error status: " + status.getStatusCode() + ", error info: " + entity + "\nqueries: " + queries, response.getStatusInfo());
-            }
-            LOG.debug("Received {}", status.getStatusCode());
-            return QueryResponse.from(queryList, entity);
-        } finally {
-            response.close();
-        }
-    }
-
-    public Optional<Keyspace> keyspace(String keyspace) throws GraknClientException {
-        URI fullURI = UriBuilder.fromUri(uri.toURI())
-                .path(REST.resolveTemplate(REST.WebPath.KB_KEYSPACE, keyspace))
-                .build();
-        ClientResponse response = client.resource(fullURI)
-                .accept(APPLICATION_JSON)
-                .get(ClientResponse.class);
-        Response.StatusType status = response.getStatusInfo();
-        LOG.debug("Received {}", status.getStatusCode());
-        if (status.getStatusCode() == Status.NOT_FOUND.getStatusCode()) {
-            return Optional.empty();
-        }
-        String entity = response.getEntity(String.class);
-        if (!status.getFamily().equals(Family.SUCCESSFUL)) {
-            throw new GraknClientException("Failed keyspace. Error status: " + status.getStatusCode() + ", error info: " + entity, response.getStatusInfo());
-        }
-        response.close();
-        return Optional.of(Keyspace.of(keyspace));
-    }
+    Optional<Keyspace> keyspace(String keyspace) throws GraknClientException;
 }
