+++ grakn-graql-shell/src/main/java/ai/grakn/graql/GraqlClient.java
@@ -69,7 +69,7 @@ class GraqlClient {
     public BatchExecutorClient loaderClient(SimpleURI uri) {
         return BatchExecutorClient.newBuilder()
                 .threadPoolCoreSize(Runtime.getRuntime().availableProcessors() * 8)
-                .taskClient(new GraknClient(uri))
+                .taskClient(GraknClient.of(uri))
                 .maxRetries(DEFAULT_MAX_RETRY)
                 .build();
     }
